// src/interfaces/BankInterface.tsx
// Generated by https://quicktype.io

export interface ResponseBank {
    ok:   boolean;
    msg:  string;
    data: DataBankInterface;
}

export interface DataBankInterface {
    id:              number;
    name_bank:       string;
    type_account_id: number;
    name_onwner:     string;
    driver_id:       number;
    number_account:  string;
    email:           string;
    identification:  string;
}

export interface BankInterface {
    name_bank:       string;
    type_account_id: number;
    name_onwner:     string;
    number_account:  string;
    email:           string;
    identification:  string;
}
// Generated by https://quicktype.io

export interface ResponseBankingParamsByIDDriver {
    ok:   boolean;
    msg:  string;
    data: DataBankIdDriver;
}

export interface DataBankIdDriver {
    id:              number;
    name_bank:       string;
    type_account_id: number;
    name_onwner:     string;
    driver_id:       number;
    number_account:  string;
    email:           string;
    identification:  string;
}